feat: add sleepsort and tests
Came across the sleepsort algorithm in a blog post and almost couldn't
believe that it actually worked. Implemented it myself just to prove
that it could be. Also added some testing to make sure everything works
correctly. The generator command has more granular control now, which is
useful for testing, especially sleepsort.